refactor: se separaron los controladores para un mejor manejo de ellos.
- Se creó el controlador ChangeStatusController para manejar la aprobación y rechazo de solicitudes de cuenta.
- Se implementó el controlador CreateAccountController para crear y actualizar cuentas en Keycloak y Moodle.
- Se añadió el controlador ListAccountRequestsController para listar solicitudes de cuenta por curso.
- Se implementó el controlador RequestAccountController para gestionar las solicitudes de cuenta.
- Se actualizaron las rutas para utilizar los nuevos controladores.
