+++ app/modules/auth/routes.py
@@ -1,6 +1,9 @@
-from fastapi import APIRouter, Depends, Query, HTTPException, Form, Request
+from fastapi import APIRouter, Depends, Query
 from app.core.database import get_db
-from .controllers import AuthController
+from .controllers.request_account import RequestAccountController
+from .controllers.list_account_requests import ListAccountRequestsController
+from .controllers.change_status import ChangeStatusController
+from .controllers.create_account import CreateAccountController
 from .schema import AccountRequestSchema, ConfirmAccountSchema, CreateAccountSchema
 from app.modules.auth.services.email_service import EmailService
 #pip install python-multipart
@@ -10,7 +13,7 @@ router = APIRouter(prefix="/auth", tags=["auth"])
 # Crear solicitud de cuenta
 @router.post("/request-account", summary="Crear una solicitud de cuenta")
 async def request_account(body_info: AccountRequestSchema, db=Depends(get_db)):
-    result = AuthController.request_account(data=body_info, db=db)
+    result = RequestAccountController.request_account(data=body_info, db=db)
     return {"success": True, "data": result}
 
 # Listar solicitudes por curso
@@ -24,12 +27,12 @@ async def list_accounts_requests(
     db=Depends(get_db)
 ):
     """Lista todas las solicitudes de cuenta filtradas por curso."""
-    return AuthController.list_accounts_requests(db=db, course_id=course_id)
+    return ListAccountRequestsController.list_accounts_requests(db=db, course_id=course_id)
 
 # Aprobar o rechazar solicitud y enviar correo
 @router.patch("/confirm-account", summary="Aprobar o rechazar solicitud y enviar correo")
 async def confirm_account(body_info: ConfirmAccountSchema, db=Depends(get_db)):
-    result = await AuthController.confirm_account(data=body_info, db=db)
+    result = await ChangeStatusController.change_status(data=body_info, db=db)
     return {"success": True, "data": result}
 
 
@@ -38,7 +41,7 @@ async def confirm_account(body_info: ConfirmAccountSchema, db=Depends(get_db)):
 @router.post("/create-account", summary="Crear cuenta en Keycloak y Moodle")
 async def create_account(body_info: CreateAccountSchema, db=Depends(get_db)):
     """Crea una cuenta en Keycloak y Moodle (administrador)."""
-    return await AuthController.create_account(data=body_info, db=db)
+    return await CreateAccountController.create_account(data=body_info, db=db)
 # Confirmar datos token
 @router.get("/confirmacion", summary="Confirmar email con token")
 def confirm_email(token: str):
